Dictionaries
For this exercise I went a bit further than the specification, just for fun.

This program always a user to add a book to a dictionary and categorise it by its genre. The user can also view the "library" which outputs the contents of the dictionary as a table of titles and quantities for each genre.