Check if a number is in a list
Generates a list of a pre-defined amount of unique random numbers. Tells the user how many numbers are in the list, in what range the numbers will be, and how many attempts the user has to guess the full list of numbers. then prompts the user for numbers until all attempts have been used.

To end the game the user either guesses all the numbers or they run out of guesses