Count vowels in a string
Prompts user for a string of text. Check each character of the text to see if it is in the list of vowels. Updates a count variables if the character is a vowel and then prints prints the value of the count variable to tell the user how many vowels are in the inputted text.